+++ src/Effects/Lift.hs
@@ -9,18 +9,34 @@
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE AllowAmbiguousTypes #-}
 
-module Effects.Lift where
+{- | For lifting arbitrary monadic computations into an algebraic effect setting
+-}
 
-import Prog
+module Effects.Lift (
+    Lift(..)
+  , lift
+  , liftM
+  , handleLift) where
 
--- ||| Lift effect for lifting an arbitrary monad `m` into the effect `Lift m`
+import Prog ( call, Member(prj), LastMember, Prog(..) )
+import Model (Model(..))
+import Data.Function (fix)
+
+-- | Lift effect
 newtype Lift m a = Lift (m a)
 
+-- | Lift a monadic computation @m a@ into the effect @Lift m@
 lift :: (LastMember (Lift m) es) => m a -> Prog es a
 lift = call . Lift
 
+-- | Lift a monadic computation @m a@ into the effect @Lift m@ in a model
+liftM :: (Member (Lift m) es) => m a -> Model env es a
+liftM op = Model (call (Lift op))
+
+-- | Handle @Lift m@ as the last effect
 handleLift :: forall m w. Monad m => Prog '[Lift m] w -> m w
-handleLift (Val x) = pure x
+handleLift (Val x) = return x
 handleLift (Op u q) = case prj u of
      Just (Lift m) -> m >>= handleLift . q
      Nothing -> error "Impossible: Nothing cannot occur"
+